+++ demo/docs/previews/code/DialogPreview.tsx
@@ -1,0 +1,5 @@
+import { DialogExample } from "../examples/dialog.example";
+
+export function DialogPreview() {
+  return <DialogExample />;
+}