+++ src/components/dialog/index.ts
@@ -1,0 +1,20 @@
+export {
+  Dialog,
+  DialogClose,
+  DialogContent,
+  DialogDescription,
+  DialogOverlay,
+  DialogPortal,
+  DialogTitle,
+  DialogTrigger,
+} from "./Dialog";
+export type {
+  DialogContentClassNames,
+  DialogContentProps,
+  DialogDescriptionClassNames,
+  DialogDescriptionProps,
+  DialogOverlayClassNames,
+  DialogOverlayProps,
+  DialogTitleClassNames,
+  DialogTitleProps,
+} from "../../types/dialog";